nlp-backend/requirements.txt
connexion==2.4.0
Flask==1.1.1
flask-cors
flask_restful==0.3.7
gunicorn==20.0.4
openapi-spec-validator==0.2.8
pandas==0.25.3
requests==2.22.0
SPARQLWrapper==1.8.4
swagger-ui-bundle==0.0.6
tqdm==4.38
spacy==2.3
geopy
paho-mqtt==1.5.1
nltk==3.6.2
pydantic==1.7